requirements.txt
streamlit
python-dotenv
PyPDF2
langchain-core
langchain-community
transformers
torch
